requirements.txt
alembic==1.4.2
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.1
decorator==4.4.2
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.5.1
nose==1.3.7
pbr==5.4.5
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.16
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
Tempita==0.5.2
urllib3==1.25.9
visitor==0.1.3
webargs==6.0.0
Werkzeug==1.0.1
WTForms==2.3.1
